+++ grace-web-mvc/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
     api project(":grace-gsp")
     api project(":grace-web")
-    api project(":grace-web-url-mappings")
 
     compileOnlyApi libs.jakarta.servlet
