+++ grace-web-mvc/src/main/groovy/org/grails/web/errors/GrailsExceptionResolver.java
@@ -15,24 +15,19 @@
  */
 package org.grails.web.errors;
 
-import java.io.IOException;
 import java.util.Collections;
 import java.util.Enumeration;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import org.codehaus.groovy.control.CompilationFailedException;
 import org.codehaus.groovy.runtime.InvokerInvocationException;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.web.context.ServletContextAware;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.View;
-import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.handler.SimpleMappingExceptionResolver;
 
 import grails.config.Config;
@@ -40,15 +35,8 @@ import grails.config.Settings;
 import grails.core.GrailsApplication;
 import grails.core.support.GrailsApplicationAware;
 import grails.util.Environment;
-import grails.web.mapping.UrlMappingInfo;
-import grails.web.mapping.UrlMappingsHolder;
-import grails.web.mapping.exceptions.UrlMappingException;
-
-import org.grails.core.exceptions.GrailsRuntimeException;
 import org.grails.exceptions.ExceptionUtils;
 import org.grails.exceptions.reporting.StackTraceFilterer;
-import org.grails.web.mapping.DefaultUrlMappingInfo;
-import org.grails.web.mapping.UrlMappingUtils;
 import org.grails.web.servlet.mvc.exceptions.GrailsMVCException;
 import org.grails.web.util.GrailsApplicationAttributes;
 import org.grails.web.util.WebUtils;
@@ -59,7 +47,7 @@ import org.grails.web.util.WebUtils;
  * @author Graeme Rocher
  * @author Michael Yan
  */
-@SuppressWarnings({ "rawtypes", "unchecked" })
+@SuppressWarnings({ "unchecked" })
 public class GrailsExceptionResolver extends SimpleMappingExceptionResolver implements ServletContextAware, GrailsApplicationAware {
 
     public static final String EXCEPTION_ATTRIBUTE = WebUtils.EXCEPTION_ATTRIBUTE;
@@ -87,11 +75,6 @@ public class GrailsExceptionResolver extends SimpleMappingExceptionResolver impl
 
         logStackTrace(ex, request);
 
-        UrlMappingsHolder urlMappings = lookupUrlMappings();
-        if (urlMappings != null) {
-            mv = resolveViewOrForward(ex, urlMappings, request, response, mv);
-        }
-
         return mv;
     }
 
@@ -117,6 +100,7 @@ public class GrailsExceptionResolver extends SimpleMappingExceptionResolver impl
 
     /**
      * Obtains the root cause of the given exception
+     *
      * @param ex The exception
      * @return The root cause
      */
@@ -143,106 +127,6 @@ public class GrailsExceptionResolver extends SimpleMappingExceptionResolver impl
         GrailsWrappedRuntimeException gwre = new GrailsWrappedRuntimeException(this.servletContext, e);
         mv.addObject(WebUtils.ERROR_EXCEPTION_ATTRIBUTE, gwre);
         mv.addObject(WebUtils.EXCEPTION_ATTRIBUTE, gwre);
-    }
-
-    protected UrlMappingsHolder lookupUrlMappings() {
-        try {
-            return UrlMappingUtils.lookupUrlMappings(this.servletContext);
-        }
-        catch (Exception ignored) {
-            // ignore, no app ctx in this case.
-            return null;
-        }
-    }
-
-    Map extractRequestParamsWithUrlMappingHolder(UrlMappingsHolder urlMappings, HttpServletRequest request) {
-        Map params = new HashMap();
-        try {
-            UrlMappingInfo requestInfo = urlMappings.match(request.getRequestURI());
-            if (requestInfo != null) {
-                params.putAll(UrlMappingUtils.findAllParamsNotInUrlMappingKeywords(requestInfo.getParameters()));
-            }
-        }
-        catch (UrlMappingException ulrMappingException) {
-            logger.debug("Could not find urlMapping which matches: " + request.getRequestURI());
-        }
-        return params;
-    }
-
-    protected ModelAndView resolveViewOrForward(Exception ex, UrlMappingsHolder urlMappings, HttpServletRequest request,
-            HttpServletResponse response, ModelAndView mv) {
-
-        UrlMappingInfo info = matchStatusCode(ex, urlMappings);
-
-        if (info != null) {
-            Map params = extractRequestParamsWithUrlMappingHolder(urlMappings, request);
-            if (params != null && !params.isEmpty()) {
-                Map infoParams = info.getParameters();
-                if (infoParams != null) {
-                    params.putAll(info.getParameters());
-                }
-                info = new DefaultUrlMappingInfo(info, params, this.grailsApplication);
-            }
-        }
-
-        try {
-            if (info != null && info.getViewName() != null) {
-                resolveView(request, info, mv);
-            }
-            else if (info != null && info.getControllerName() != null) {
-                String uri = determineUri(request);
-                if (!response.isCommitted()) {
-                    forwardRequest(info, request, response, mv, uri);
-                    // return an empty ModelAndView since the error handler has been processed
-                    return new ModelAndView();
-                }
-            }
-            return mv;
-        }
-        catch (Exception e) {
-            logger.error("Unable to render errors view: " + e.getMessage(), e);
-            throw new GrailsRuntimeException(e);
-        }
-    }
-
-    protected void forwardRequest(UrlMappingInfo info, HttpServletRequest request, HttpServletResponse response,
-            ModelAndView mv, String uri) throws ServletException, IOException {
-
-        info.configure(WebUtils.retrieveGrailsWebRequest());
-        String forwardUrl = UrlMappingUtils.forwardRequestForUrlMappingInfo(
-                request, response, info, mv.getModel(), true);
-        if (logger.isDebugEnabled()) {
-            logger.debug("Matched URI [" + uri + "] to URL mapping [" + info +
-                    "], forwarding to [" + forwardUrl + "] with response [" + response.getClass() + "]");
-        }
-    }
-
-    protected String determineUri(HttpServletRequest request) {
-        String uri = (String) request.getAttribute(WebUtils.FORWARD_REQUEST_URI_ATTRIBUTE);
-        if (uri == null) {
-            uri = request.getRequestURI();
-        }
-        return uri;
-    }
-
-    protected void resolveView(HttpServletRequest request, UrlMappingInfo info, ModelAndView mv) throws Exception {
-        ViewResolver viewResolver = WebUtils.lookupViewResolver(this.servletContext);
-        View v = UrlMappingUtils.resolveView(request, info, info.getViewName(), viewResolver);
-        if (v != null) {
-            mv.setView(v);
-        }
-    }
-
-    protected UrlMappingInfo matchStatusCode(Exception ex, UrlMappingsHolder urlMappings) {
-        UrlMappingInfo info = urlMappings.matchStatusCode(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, ex);
-        if (info == null) {
-            info = urlMappings.matchStatusCode(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,
-                    getRootCause(ex));
-        }
-        if (info == null) {
-            info = urlMappings.matchStatusCode(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-        }
-        return info;
     }
 
     protected void logStackTrace(Exception e, HttpServletRequest request) {
