+++ grace-web-url-mappings/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     api project(":grace-api")
     api project(":grace-util")
     api project(":grace-web")
+    api project(":grace-web-mvc")
 
     api(libs.grace.datastore.gorm.validation)
